requirements.txt
fastapi
uvicorn
pydantic
requests
psycopg2-binary
